Major UI updates and changes addition of manual modbus device control
